Index (lines metadata) moved from RAM to file
Now building an index is inefficient, but it will be improved.

Moving index from RAM to file has two main effects:
- now there is no limit for a line count. Before we had to fit index in RAM,
  so the limit for line count was about #RAM / (size of 1 line metadata). For
  degenerated files (only empty lines) we could handle only files with size
  < 1 GB.
- get command will be aprox. 2 times slower than before
	before: ~1 disk read
	now: ~2 disk reads

+++ line-server/src/main/java/com/stefanski/lineserver/LineServerProtocol.java
@@ -38,9 +38,9 @@ class LineServerProtocol {
             throw new IllegalArgumentException("Invalid format of get command: " + input);
         }
 
-        int lineNr;
+        long lineNr;
         try {
-            lineNr = Integer.valueOf(tokens[1]);
+            lineNr = Long.valueOf(tokens[1]);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("Cannot parse line number: " + tokens[1], e);
         }
